Solution to Calculate and Print the Average of Three Numbers

+++ Calculate and Print the Average of Three Numbers@@6788d5012db25efec946f944/code.c
@@ -3,7 +3,7 @@ int main(){
     int k;
     int j;
     int m;
-    float l;
+    int l;
     l=(k+j+m)/3;
     printf("%.2f",l);
     return 0;
